Add files via upload
 Even if the  locale was consistent with the input filename, it would still
be like this problem  ,that is,non-ascii  filenames get truncated. After I
debugged it, I found that the* len* variable for the third parameter of the
`wcstombs` function(Line 99 in basename.c and Line 142 in dirname.c) has
the value returned by `mbstowcs`. However,the *len* value returned by the
`mbstowcs` function refers to the number of characters, while the third
parameter  of the wcstombs function refers to the size of the target buffer
in bytes.I think it is best to use wcstombs (NULL, wcstr, 0) to get the size correctly.